pyproject.toml
[tool.poetry]
name = "nasa-backend"
version = "0.1.0"
description = ""
authors = ["Jean-Shan Chou <jschou@cs.nctu.edu.tw>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.0"
pandas = "^2.1.1"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.21"
psycopg2-binary = "^2.9.9"
geopandas = "^0.14.0"
geojson = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

